% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{available_contrasts}
\alias{available_contrasts}
\title{List Available Causal Contrasts}
\usage{
available_contrasts()
}
\value{
A data.frame describing each supported contrast.
}
\description{
List Available Causal Contrasts
}
