% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aipw_estimate.R
\name{print.aipw_result}
\alias{print.aipw_result}
\title{Print method for aipw_result}
\usage{
\method{print}{aipw_result}(x, digits = 4, ...)
}
\arguments{
\item{x}{An aipw_result object.}

\item{digits}{Number of digits. Default 4.}

\item{...}{Ignored.}
}
\description{
Print method for aipw_result
}
